// src/lib/components/navbar-buttons/CurrentSelectionButton/CurrentSelectionPanel/types/selection-group.ts
/*
 * AtroCore Software
 *
 * This source file is available under GNU General Public License version 3 (GPLv3).
 * Full copyright and license information is available in LICENSE.txt, located in the root directory.
 *
 * @copyright  Copyright (c) AtroCore GmbH (https://www.atrocore.com)
 * @license    GPLv3 (https://www.gnu.org/licenses/)
 */

import type SelectionItem from './selection-item';

type SelectionGroup = {
    key: string;
    collection: SelectionItem[];
    rowList: string[];
    icon?: string | null;
};

export default SelectionGroup;